Updated to generate plans as we work through the range stream
