+++ warehouse/query-core/src/main/java/datawave/query/planner/DatePartitionedQueryIterable.java
@@ -1,30 +1,40 @@
 package datawave.query.planner;
 
 import java.io.IOException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.Iterator;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
+import org.apache.commons.lang3.tuple.Pair;
+import org.apache.log4j.Logger;
+
+import datawave.core.common.logging.ThreadConfigurableLogger;
 import datawave.core.query.configuration.QueryData;
 import datawave.query.CloseableIterable;
+import datawave.query.config.ShardQueryConfiguration;
+import datawave.query.exceptions.DatawaveFatalQueryException;
+import datawave.query.exceptions.DatawaveQueryException;
 
 /**
  * Implementation of {@link CloseableIterable} intended to be used by {@link DatePartitionedQueryPlanner}. This iterable
  */
 public class DatePartitionedQueryIterable implements CloseableIterable<QueryData> {
 
-    private final List<CloseableIterable<QueryData>> iterables = new ArrayList<>();
+    private static final Logger log = ThreadConfigurableLogger.getLogger(DatePartitionedQueryIterable.class);
+    private final List<SubPlanCallable> subPlanGenerators;
+    private final ShardQueryConfiguration planningConfig;
+    private Iter iterator;
 
-    /**
-     * Add an iterable to this {@link DatePartitionedQueryIterable}.
-     *
-     * @param iterable
-     *            the iterable to add
-     */
-    public void addIterable(CloseableIterable<QueryData> iterable) {
-        if (iterable != null) {
-            iterables.add(iterable);
-        }
+    public DatePartitionedQueryIterable(List<SubPlanCallable> subPlansGenerators, ShardQueryConfiguration planningConfig)
+                    throws DatawaveQueryException, DatawaveFatalQueryException {
+        this.subPlanGenerators = subPlansGenerators;
+        this.planningConfig = planningConfig;
+        this.iterator = new Iter();
     }
 
     /**
@@ -35,10 +45,9 @@ public class DatePartitionedQueryIterable implements CloseableIterable<QueryData
      */
     @Override
     public void close() throws IOException {
-        for (CloseableIterable<QueryData> iterable : iterables) {
-            iterable.close();
+        if (this.iterator != null) {
+            this.iterator.close();
         }
-        iterables.clear();
     }
 
     /**
@@ -48,19 +57,55 @@ public class DatePartitionedQueryIterable implements CloseableIterable<QueryData
      */
     @Override
     public Iterator<QueryData> iterator() {
-        return new Iter();
+        return this.iterator;
     }
 
     /**
-     * Iterator implementation that provides the ability to iterate over each {@link QueryData} of the iterables in {@link #iterables}.
+     * Get the planned script
+     *
+     * @return the planned script
+     */
+    public String getPlannedScript() {
+        return this.iterator.getPlannedScript();
+    }
+
+    /**
+     * Iterator implementation that provides the ability to iterate over each {@link QueryData} of the iterables generated by the {@link SubPlanCallable}.
      */
     private class Iter implements Iterator<QueryData> {
 
-        // Iterator that traverses over the iterables.
-        private final Iterator<CloseableIterable<QueryData>> iterableIterator = iterables.iterator();
+        private final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd");
+
+        // the plans
+        private final Set<String> plans = new LinkedHashSet<>();
+
+        // the exceptions
+        private List<DatawaveQueryException> exceptions = new ArrayList<>();
+
+        // Iterator that traverses over the plans
+        private final Iterator<SubPlanCallable> callableIterator = subPlanGenerators.iterator();
+
+        // The closable iterable
+        private CloseableIterable<QueryData> currentSubIterable = null;
 
         // The current sub iterator.
         private Iterator<QueryData> currentSubIterator = null;
+
+        // Construct the iterator which will in turn find an available query data stream
+        public Iter() throws DatawaveQueryException, DatawaveFatalQueryException {
+            // Lets make sure we have at least one valid query data stream
+            seekToNextAvailableQueryData();
+
+            // if every plan failed, then pass an exception up
+            evaluateFailures();
+        }
+
+        public void close() throws IOException {
+            if (currentSubIterable != null) {
+                this.currentSubIterable.close();
+                this.currentSubIterable = null;
+            }
+        }
 
         @Override
         public boolean hasNext() {
@@ -78,23 +123,121 @@ public class DatePartitionedQueryIterable implements CloseableIterable<QueryData
          */
         private void seekToNextAvailableQueryData() {
             // If the current sub iterator is null, attempt to get the next available iterator, or return early if there are no more iterators.
-            if (currentSubIterator == null) {
-                if (iterableIterator.hasNext()) {
-                    currentSubIterator = iterableIterator.next().iterator();
-                } else {
-                    return;
-                }
+            while ((currentSubIterator == null || !currentSubIterator.hasNext()) && callableIterator.hasNext()) {
+                populateNextIterable();
             }
-            // If the current sub iterator does not have any more elements remaining, move to the next sub iterator that does have elements.
-            if (!currentSubIterator.hasNext()) {
-                while (iterableIterator.hasNext()) {
-                    // We must ensure we only ever call iterator() once on each sub-iterator.
-                    currentSubIterator = iterableIterator.next().iterator();
-                    if (currentSubIterator.hasNext()) {
-                        return;
+        }
+
+        private void populateNextIterable() {
+            if (callableIterator.hasNext()) {
+                SubPlanCallable callable = callableIterator.next();
+                Map.Entry<Pair<Date,Date>,Set<String>> dateRange = callable.getDateRange();
+                String subBeginDate = dateFormat.format(dateRange.getKey().getLeft());
+                String subEndDate = dateFormat.format(dateRange.getKey().getRight());
+
+                try {
+                    currentSubIterable = callable.call();
+                    // may be null if only generating a plan
+                    if (currentSubIterable != null) {
+                        currentSubIterator = currentSubIterable.iterator();
                     }
+
+                    if (log.isDebugEnabled()) {
+                        log.debug("Query string for config of sub-plan against date range (" + subBeginDate + "-" + subEndDate + ") with unindexed fields "
+                                        + dateRange.getValue() + ": " + callable.getSubPlanConfig().getQueryString());
+                    }
+                } catch (Exception e) {
+                    String msg = "Exception occurred when processing sub-plan against date range (" + subBeginDate + "-" + subEndDate + ")";
+                    log.warn(msg, e);
+                    exceptions.add(new DatawaveQueryException(msg, e.getCause()));
+                } finally {
+                    ShardQueryConfiguration subPlanConfig = callable.getSubPlanConfig();
+
+                    // append the new timers for logging at the end
+                    planningConfig.appendTimers(subPlanConfig.getTimers());
+
+                    // Add to the set of plans
+                    plans.add(subPlanConfig.getQueryString());
                 }
             }
         }
+
+        /**
+         * Get the planned script to represent a concatenation of the planned scripts from all sub-plans. If we have not gathered all of the subplans, then
+         * simply return the original plan
+         */
+        public String getPlannedScript() {
+            String plannedScript;
+            if (plans.size() == subPlanGenerators.size()) {
+                if (plans.isEmpty()) {
+                    plannedScript = planningConfig.getQueryString();
+                } else if (plans.size() == 1) {
+                    plannedScript = plans.iterator().next();
+                } else {
+                    StringBuilder sb = new StringBuilder();
+                    int i = 0;
+                    for (String plan : plans) {
+                        if (sb.length() > 0) {
+                            sb.append(" || ");
+                        }
+                        sb.append("((plan = ").append(++i).append(") && (").append(plan).append("))");
+                    }
+                    plannedScript = sb.toString();
+                }
+            } else {
+                plannedScript = planningConfig.getQueryString();
+            }
+            return plannedScript;
+        }
+
+        /**
+         * Determine if we are failing the query entirely. This can happen if every subplan failed to execute.
+         *
+         * @throws DatawaveQueryException
+         * @throws DatawaveFatalQueryException
+         */
+        private void evaluateFailures() throws DatawaveQueryException, DatawaveFatalQueryException {
+            if (exceptions.size() == subPlanGenerators.size()) {
+                if (exceptions.size() == 1) {
+                    Throwable e = unwrapException(exceptions.get(0));
+                    if (e instanceof RuntimeException) {
+                        throw (RuntimeException) e;
+                    } else {
+                        throw (DatawaveQueryException) e;
+                    }
+                } else {
+                    DatawaveFatalQueryException e = new DatawaveFatalQueryException("Query failed creation for " + planningConfig.getQuery().getId());
+                    for (Throwable reason : exceptions) {
+                        e.addSuppressed(reason);
+                    }
+                    throw e;
+                }
+            }
+        }
+
+        /**
+         * In the case were we have only 1 exception, we want to throw the underlying exception to be backward compatible which will be the lowest level
+         * DatawaveQueryException or first DatawaveFatalQueryException
+         *
+         * @param e
+         *            The exception being unwrapped
+         * @return DatawaveQueryException or DatawaveFatalQueryException
+         */
+        private Throwable unwrapException(DatawaveQueryException e) {
+            Throwable toThrow = e;
+            Throwable test = e;
+            while (test.getCause() != null && test.getCause() != test) {
+                test = test.getCause();
+                if (test instanceof DatawaveFatalQueryException) {
+                    return test;
+                }
+                if (test instanceof DatawaveQueryException) {
+                    toThrow = test;
+                }
+            }
+            return toThrow;
+        }
+
     }
+
 }
